adds re-usable hook
